fix: update acir.hpp and brillig_vm test bit sizes

+++ barretenberg/cpp/src/barretenberg/dsl/acir_format/serde/acir.hpp
@@ -1234,7 +1234,7 @@ struct AssertionPayload {
     };
 
     struct Dynamic {
-        std::tuple<uint64_t, std::vector<Program::ExpressionOrMemory>> value;
+        std::tuple<uint32_t, std::vector<Program::ExpressionOrMemory>> value;
 
         friend bool operator==(const Dynamic&, const Dynamic&);
         std::vector<uint8_t> bincodeSerialize() const;
